fix(cpu/matmul): widen GEMV weight row offsets to long (#429)
Port of upstream PR kkokosa/dotLLM#431 onto dev, extended to the two GEMV
kernels that exist only on this fork.

Upstream sites (MatMul.cs): ComputeRows `weightsQ8 + row * rowBytes` across the
AVX-512+VNNI / AVX-512 / AVX2 / scalar tiers, and GemvF16 `weightsHalf + row * k`.
The Q8_0 stride is in bytes (~1.0625/weight) so the product wraps at 2 GB of
tensor: Llama 3.1 405B's LM head (vocab 128256, rowBytes 17408) reaches
2,232,680,448, over int.MaxValue by ~85 MB, so its last 4,894 rows read from a
wrapped negative offset. GemvF16 counts elements and stays ~2.2% under on the
same model.

Fork-only sites, not covered by the upstream issue or PR:

- MatMulVnni.cs `weightsQ8 + row * rowBytes` (2 sites) — identical byte stride to
  ComputeRows, so this is EQUALLY REACHABLE on the same model. It also matters
  more in practice: this is the path modern VNNI-capable hardware dispatches to.
- MatMulMxfp4.cs `weights + row * rowBytes` (2 sites) — same shape, but MXFP4's
  ~0.53 bytes/weight puts that model at 1,116,340,224, roughly half the limit.
  Widened for consistency; not currently reachable.

No performance cost — the JIT already sign-extended after the multiply; the cast
moves the widening ahead of it, leaving the 4-row fast path one instruction
shorter. See kkokosa/dotLLM#431 for the codegen diff and the (inconclusive)
timing data.

Reported-by: unsafePtr

+++ src/DotLLM.Cpu/Kernels/MatMulMxfp4.cs
@@ -192,12 +192,12 @@ public static unsafe partial class MatMul
         if (Avx2.IsSupported && Ssse3.IsSupported)
         {
             for (int row = 0; row < m; row++)
-                result[row] = VecDotMxfp4Q8_0Avx2(weights + row * rowBytes, xQ8, blockCount);
+                result[row] = VecDotMxfp4Q8_0Avx2(weights + (long)row * rowBytes, xQ8, blockCount);
         }
         else
         {
             for (int row = 0; row < m; row++)
-                result[row] = VecDotMxfp4Q8_0Scalar(weights + row * rowBytes, xQ8, blockCount);
+                result[row] = VecDotMxfp4Q8_0Scalar(weights + (long)row * rowBytes, xQ8, blockCount);
         }
     }
